# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MinRegistry management API",
    "description": "Management API of MinRegistry. All endpoints except this document require an admin session (GitHub OAuth); mutating requests must send `X-Requested-With: XMLHttpRequest`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/principals": {
      "get": {
        "tags": [
          "principals"
        ],
        "summary": "GitHub admins (from configuration) and all principals.",
        "operationId": "list_principals",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "principals"
        ],
        "summary": "Creates a token-only identity (no web login).",
        "operationId": "create_identity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalSummary"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Name already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/principals/{id}": {
      "get": {
        "tags": [
          "principals"
        ],
        "summary": "A principal with its tokens and repository grants.",
        "operationId": "get_principal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "principals"
        ],
        "summary": "Enables/disables an identity or renames its display name.",
        "operationId": "update_principal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrincipalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalSummary"
                }
              }
            }
          },
          "400": {
            "description": "GitHub admins are managed through configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/principals/{id}/tokens": {
      "get": {
        "tags": [
          "principals"
        ],
        "summary": "Tokens of a principal (secrets are never returned).",
        "operationId": "list_tokens",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenSummary"
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "principals"
        ],
        "summary": "Issues a token. The secret is returned once. Admins may mint tokens for\nidentities and for their own GitHub principal only.",
        "operationId": "create_token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Another admin's GitHub principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/principals/{id}/tokens/{token_id}": {
      "delete": {
        "tags": [
          "principals"
        ],
        "summary": "Revokes a token immediately.",
        "operationId": "revoke_token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token_id",
            "in": "path",
            "description": "Token id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIdentityRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Lower-case letters, digits, `.`, `_`, `-`; used as the `docker login` username."
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateTokenRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional RFC 3339 expiry; tokens never expire by default."
          }
        }
      },
      "CreateTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "secret",
          "username"
        ],
        "properties": {
          "token": {
            "$ref": "#/components/schemas/TokenSummary"
          },
          "secret": {
            "type": "string",
            "description": "The secret. Shown only in this response; it cannot be retrieved later."
          },
          "username": {
            "type": "string",
            "description": "Username to pair with the secret (`docker login -u`)."
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable code, e.g. `not_found`, `unauthenticated`, `csrf`."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "PermissionLevel": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "owner"
        ]
      },
...
      "PrincipalDetail": {
        "type": "object",
        "required": [
          "principal",
          "tokens",
          "permissions"
        ],
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/PrincipalSummary"
          },
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenSummary"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrincipalPermissionSummary"
            }
          }
        }
      },
      "PrincipalKind": {
        "type": "string",
        "enum": [
          "github",
          "identity"
        ]
      },
      "PrincipalListResponse": {
        "type": "object",
        "required": [
          "admin_logins",
          "principals"
        ],
        "properties": {
          "admin_logins": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "GitHub logins allowed to sign in (from configuration, read-only)."
          },
          "principals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrincipalSummary"
            }
          }
        }
      },
      "PrincipalPermissionSummary": {
        "type": "object",
        "required": [
          "repository",
          "level",
          "granted_at"
        ],
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryRefSummary"
          },
          "level": {
            "$ref": "#/components/schemas/PermissionLevel"
          },
          "granted_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "granted_at": {
            "type": "string"
          }
        }
      },
...
      "PrincipalSummary": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "name",
          "display_name",
          "enabled",
          "is_admin",
          "active",
          "active_token_count",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PrincipalKind"
          },
          "name": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "is_admin": {
            "type": "boolean",
            "description": "GitHub principal currently listed in `MINREGISTRY_ADMIN_GITHUB_LOGINS`."
          },
          "active": {
            "type": "boolean",
            "description": "Whether the principal can authenticate right now."
          },
          "active_token_count": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
...
      "RepositoryRefSummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "TokenStatus": {
        "type": "string",
        "enum": [
          "active",
          "expired",
          "revoked"
        ]
      },
      "TokenSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "First characters of the secret, to recognize it."
          },
          "status": {
            "$ref": "#/components/schemas/TokenStatus"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UpdatePrincipalRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "minregistry_session"
      }
    }
  },
...
}